refactor(firmware): extract FirmwareManager to src/firmware/firmware_manager.py
Move 2326-line FirmwareManager class to standalone module using dependency
injection pattern consistent with other extracted firmware classes.

Closes #203
